#37 refactor : 상품 별 재고 합 및 비율 기능 수정

+++ src/main/java/com/hmdandelion/project_1410002/inventory/domian/repository/stock/StockRepo.java
@@ -19,7 +19,7 @@ public interface StockRepo extends JpaRepository<Stock,Long>, StockRepoCustom {
 @Query("SELECT SUM(s.quantity) FROM Stock s WHERE s.isDelete=false")
     Integer getAccumulateStock();
 @Query("SELECT SUM(s.quantity) FROM Stock s WHERE s.product.productCode = :productCode AND s.isDelete=false")
-    Integer getAccumulateStockByProductCode(Integer productCode);
+    Long getAccumulateStockByProductCode(Long productCode);
 
     List<Stock> findByProductProductCodeAndIsDelete(Long productCode,Boolean isDelete);
 }
